Committing maintenance page.
Committing the maintenance page used when the site is offline.
Previously, without this page in the theme, the site defaulted to the
system maintenance page located in /modules/system which is incompatible
with some of the Zen theming settings, creating a broken maintenance
page. This is the Zen version, modified with the custom <div id="base">
which inserts the bottom drop-shadow image that lies under the page
content.